+++ contracts/modules/publish/PublishModule.sol
@@ -97,7 +97,7 @@ contract PublishModule is FeeModuleBase, IPublishModule, ModuleBase {
     function updatePublish(
         uint256 publishId,
         uint256 salePrice,
-        uint256 royaltyBasisPoints,
+        uint16 royaltyBasisPoints,
         uint256 amount,
         string memory name,
         string memory description,
